refactor(engine): split planetscale.go and its tests by responsibility
Move the 3,172-line pkg/engine/planetscale/planetscale.go into seven
focused files within the same package, and mirror the split in the
test file. Pure file-organization refactor — no logic, signature, or
visibility changes.

Production files:
- planetscale.go: package doc, types/consts, constructors, credential
  helpers, retry/backoff, deploy-state mapping, metadata codec,
  resume-state builders, generateBranchName, formatDeployRequestError.
- plan.go: Plan + plan-time schema helpers.
- apply.go: Apply + applyChangesToBranch + applyKeyspaceChanges(Once),
  VSchema apply helpers, diffBranchForResume, resumeApply.
- branch.go: branch lifecycle and verification (createBranch,
  waitForBranchReady, createDeployRequest, getDeployRequest,
  verifyBranchMatchesDesired*, fetch*Schema, diffKeyspace,
  verifyBranchMatchesMain).
- progress.go: Progress + Vitess-migration discovery and shard
  aggregation.
- control.go: Stop, Start, Cutover, Revert, SkipRevert, controlMeta.
- volume.go: Volume + volumeToThrottleRatio.

Test files (mirror the production split, tests sit next to the file
they exercise):
- planetscale_test.go: residual (deploy-state mapping, branch-name,
  metadata codec, retry/backoff, SplitStatements which targets
  pkg/ddl, FormatDeployRequestError).
- apply_test.go: TestApply_*, TestApplyKeyspaceChanges_* (plus the
  permanentVSchemaErrorClient helper used only by those tests).
- branch_test.go: TestDiffKeyspace_DetectsSchemaChanges.
- progress_test.go: TestAggregateShardProgress, TestValidateMigrationContext,
  TestShardLess.
- control_test.go: TestStart_Rejects/AcceptsDeferredDeploy.
- volume_test.go: TestVolumeToThrottleRatio.

Split script was hardened over PR 191's lessons: each chunk ends at
doc_start(next_func) - 1, so next-function doc comments don't leak.
Package doc lives only in planetscale.go.

Validation:
- go build ./...
- go test -count=1 -race ./pkg/engine/planetscale/... (35 tests pass)
- make test-unit (all packages pass)
- golangci-lint run --new-from-rev=origin/main ./pkg/engine/planetscale/... (0 issues)
