+++ pkg/engine/planetscale/planetscale.go
@@ -391,24 +391,17 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"maps"
 	"math/rand/v2"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	mysql "github.com/go-sql-driver/mysql"
 	ps "github.com/planetscale/planetscale-go/planetscale"
 
-	"github.com/block/spirit/pkg/statement"
-	"github.com/block/spirit/pkg/table"
 	"github.com/block/spirit/pkg/utils"
-	"golang.org/x/sync/errgroup"
 
-	"github.com/block/schemabot/pkg/ddl"
 	"github.com/block/schemabot/pkg/engine"
 	"github.com/block/schemabot/pkg/lint"
 	"github.com/block/schemabot/pkg/psclient"
@@ -664,673 +657,6 @@ func mainBranch(creds *engine.Credentials) string {
 	return "main"
 }
 
-// --- Plan ---
-
-// Plan computes the schema changes needed by diffing current schema against desired.
-// For each keyspace in the schema files, it fetches the current schema and uses
-// Spirit's PlanChanges to diff and lint in a single pass.
-func (e *Engine) Plan(ctx context.Context, req *engine.PlanRequest) (*engine.PlanResult, error) {
-	e.logger.Info("computing plan",
-		"database", req.Database,
-		"schema_files", len(req.SchemaFiles),
-	)
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	org := credOrg(req.Credentials)
-	branch := mainBranch(req.Credentials)
-
-	// Sort keyspaces for deterministic order
-	keyspaces := sortedKeyspaces(req.SchemaFiles)
-
-	// Prefer the PlanetScale schema API when safe schema changes are enabled,
-	// and use vtgate only when they are not.
-	currentSchema, err := e.fetchPlanSchema(ctx, client, org, req.Database, branch, req.Credentials, keyspaces)
-	if err != nil {
-		return nil, fmt.Errorf("fetch current schema: %w", err)
-	}
-
-	// Diff and lint per keyspace in parallel using Spirit's PlanChanges.
-	type keyspaceResult struct {
-		change     engine.SchemaChange
-		violations []engine.LintViolation
-		schemas    map[string]string // keyspace.table -> CREATE TABLE
-		hasChanges bool
-	}
-
-	var mu sync.Mutex
-	results := make(map[string]*keyspaceResult, len(keyspaces))
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(20)
-
-	for _, keyspace := range keyspaces {
-		ks := keyspace
-		g.Go(func() error {
-			ns := req.SchemaFiles[ks]
-
-			tableChanges, vschemaChanged, diffErr := e.diffKeyspace(gCtx, client, org, req.Database, branch, ks, ns, currentSchema)
-			if diffErr != nil {
-				return diffErr
-			}
-
-			sc := engine.SchemaChange{
-				Namespace:    ks,
-				Metadata:     make(map[string]string),
-				TableChanges: tableChanges,
-			}
-
-			if vschemaChanged {
-				currentVSchemaRaw := ""
-				currentVSchema, _ := client.GetKeyspaceVSchema(gCtx, &ps.GetKeyspaceVSchemaRequest{
-					Organization: org, Database: req.Database, Branch: branch, Keyspace: ks,
-				})
-				if currentVSchema != nil {
-					currentVSchemaRaw = currentVSchema.Raw
-				}
-				sc.Metadata["vschema_changed"] = "true"
-				sc.Metadata["vschema"] = VSchemaDiff(currentVSchemaRaw, ns.Files["vschema.json"])
-			}
-
-			var currentTableSchemas []table.TableSchema
-			if tables, ok := currentSchema[ks]; ok {
-				currentTableSchemas = append(currentTableSchemas, tables...)
-			}
-			desiredTableSchemas, _ := parseDesiredSchemas(ks, ns)
-			plan, _ := lint.PlanChanges(currentTableSchemas, desiredTableSchemas, nil, e.linter.SpiritConfig())
-
-			schemas := make(map[string]string)
-			if tables, ok := currentSchema[ks]; ok {
-				for _, t := range tables {
-					schemas[ks+"."+t.Name] = t.Schema
-				}
-			}
-
-			var violations []engine.LintViolation
-			if plan != nil {
-				for _, pc := range plan.Changes {
-					for _, v := range pc.Violations {
-						violations = append(violations, engine.LintViolation{
-							Table:    pc.TableName,
-							Linter:   v.Linter.Name(),
-							Message:  v.Message,
-							Severity: strings.ToLower(v.Severity.String()),
-						})
-					}
-				}
-			}
-
-			mu.Lock()
-			results[ks] = &keyspaceResult{
-				change:     sc,
-				violations: violations,
-				schemas:    schemas,
-				hasChanges: len(sc.TableChanges) > 0 || sc.Metadata["vschema_changed"] == "true",
-			}
-			mu.Unlock()
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-
-	// Collect results in deterministic keyspace order, deduplicating lint violations.
-	var changes []engine.SchemaChange
-	var lintViolations []engine.LintViolation
-	seenLint := make(map[string]bool)
-	originalSchema := make(map[string]string)
-	for _, ks := range keyspaces {
-		r := results[ks]
-		if r == nil {
-			continue
-		}
-		maps.Copy(originalSchema, r.schemas)
-		for _, v := range r.violations {
-			key := v.Table + "\x00" + v.Message
-			if !seenLint[key] {
-				seenLint[key] = true
-				lintViolations = append(lintViolations, v)
-			}
-		}
-		if r.hasChanges {
-			changes = append(changes, r.change)
-		}
-	}
-
-	if len(changes) == 0 {
-		return &engine.PlanResult{
-			PlanID:    fmt.Sprintf("plan-%d", time.Now().UnixNano()),
-			NoChanges: true,
-		}, nil
-	}
-
-	return &engine.PlanResult{
-		PlanID:         fmt.Sprintf("plan-%d", time.Now().UnixNano()),
-		Changes:        changes,
-		LintViolations: lintViolations,
-		OriginalSchema: originalSchema,
-	}, nil
-}
-
-// --- Apply ---
-
-// Apply starts executing a schema change plan.
-// Creates a PlanetScale branch, applies DDL via MySQL connection to the branch,
-// then creates and starts a deploy request.
-func (e *Engine) Apply(ctx context.Context, req *engine.ApplyRequest) (*engine.ApplyResult, error) {
-	e.logger.Info("applying plan",
-		"plan_id", req.PlanID,
-		"database", req.Database,
-	)
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	org := credOrg(req.Credentials)
-	main := mainBranch(req.Credentials)
-
-	// Check if resuming
-	if req.ResumeState != nil && req.ResumeState.Metadata != "" {
-		return e.resumeApply(ctx, client, org, req)
-	}
-
-	// emitEvent logs a lifecycle event and sends it to the caller for apply_logs recording.
-	emitEvent := func(event engine.ApplyEvent) {
-		attrs := []any{"database", req.Database}
-		for k, v := range event.Metadata {
-			attrs = append(attrs, k, v)
-		}
-		e.logger.Info(event.Message, attrs...)
-		if req.OnEvent != nil {
-			req.OnEvent(event)
-		}
-	}
-
-	// Track in-flight apply metadata for progress queries during setup.
-	migCtx := ""
-	if req.ResumeState != nil {
-		migCtx = req.ResumeState.MigrationContext
-	}
-	// persistState persists apply metadata to storage via OnStateChange for crash recovery.
-	// On first apply, migCtx is empty until the tern layer assigns one via ResumeState.
-	// persistState is a no-op in this window — if the worker crashes before Apply returns,
-	// there's no ResumeState to recover from. The tern layer handles this by retrying
-	// the full Apply on the next heartbeat recovery cycle.
-	persistState := func(meta *psMetadata) {
-		if migCtx == "" || req.OnStateChange == nil {
-			return
-		}
-		encoded, err := encodePSMetadata(meta)
-		if err != nil {
-			e.logger.Warn("failed to encode apply metadata for persistence", "error", err)
-			return
-		}
-		req.OnStateChange(&engine.ResumeState{
-			MigrationContext: migCtx,
-			Metadata:         encoded,
-		})
-	}
-
-	// Create or reuse a branch
-	existingBranch := req.Options["branch"]
-	var branchName string
-	branchStart := time.Now()
-
-	if existingBranch != "" {
-		// Reuse existing branch: wait for ready, refresh schema from main, wait again
-		branchName = existingBranch
-		if branchName == main {
-			return nil, engine.NewPermanentError("cannot reuse the %s branch: use a development branch", main)
-		}
-		persistState(&psMetadata{BranchName: branchName})
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Reusing branch %s", branchName),
-			Metadata: map[string]string{"branch": branchName},
-			NewState: state.Apply.PreparingBranch,
-		})
-
-		// Verify branch exists
-		if _, err := client.GetBranch(ctx, &ps.GetDatabaseBranchRequest{
-			Organization: org, Database: req.Database, Branch: branchName,
-		}); err != nil {
-			return nil, fmt.Errorf("branch %s not found: %w", branchName, err)
-		}
-
-		// Wait for branch to be ready (may be initializing from a prior create)
-		if err := e.waitForBranchReady(ctx, client, org, req.Database, branchName); err != nil {
-			return nil, fmt.Errorf("wait for branch %s: %w", branchName, err)
-		}
-
-		// Sync with main to pick up latest schema
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Refreshing schema for branch %s from %s", branchName, main),
-			Metadata: map[string]string{"branch": branchName},
-		})
-		if err := client.RefreshSchema(ctx, org, req.Database, branchName); err != nil {
-			return nil, fmt.Errorf("refresh schema for branch %s: %w", branchName, err)
-		}
-
-		// Wait for sync to complete
-		if err := e.waitForBranchReady(ctx, client, org, req.Database, branchName); err != nil {
-			return nil, fmt.Errorf("wait for schema refresh %s: %w", branchName, err)
-		}
-		elapsed := time.Since(branchStart).Round(time.Second)
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Branch %s schema refreshed (%s)", branchName, elapsed),
-			Metadata: map[string]string{"branch": branchName},
-			NewState: state.Apply.ApplyingBranchChanges,
-		})
-
-	} else {
-		// Create a new branch
-		branchName = generateBranchName(req.Database, req.PlanID)
-		persistState(&psMetadata{BranchName: branchName})
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Creating branch %s", branchName),
-			Metadata: map[string]string{"branch": branchName},
-			NewState: state.Apply.PreparingBranch,
-		})
-
-		_, err = e.createBranch(ctx, client, org, req.Database, branchName, main)
-		if err != nil {
-			return nil, fmt.Errorf("create branch: %w", err)
-		}
-
-		// Wait for branch to be ready
-		if err := e.waitForBranchReady(ctx, client, org, req.Database, branchName); err != nil {
-			return nil, fmt.Errorf("wait for branch: %w", err)
-		}
-		elapsed := time.Since(branchStart).Round(time.Second)
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Branch %s ready (%s)", branchName, elapsed),
-			Metadata: map[string]string{"branch": branchName},
-			NewState: state.Apply.ApplyingBranchChanges,
-		})
-	}
-
-	// Get branch credentials for MySQL access (used for DDL apply and validation).
-	pwCtx, pwCancel := context.WithTimeout(ctx, 10*time.Second)
-	defer pwCancel()
-
-	password, err := client.CreateBranchPassword(pwCtx, &ps.DatabaseBranchPasswordRequest{
-		Organization: org,
-		Database:     req.Database,
-		Branch:       branchName,
-		Role:         "admin",
-		TTL:          3600,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("create branch password: %w", err)
-	}
-
-	// For reused branches, verify the branch schema matches main. If the
-	// branch has stale DDL from a previous failed apply, RefreshSchema won't
-	// remove it — the branch will be ahead of main, producing inverted
-	// diffs (e.g., DROP COLUMN instead of ADD COLUMN).
-	// Uses MySQL to fetch the branch schema (real-time, no API staleness).
-	if existingBranch != "" {
-		keyspaces := sortedKeyspaces(req.SchemaFiles)
-		if err := e.verifyBranchMatchesMain(ctx, client, org, req.Database, branchName, main, keyspaces, req.SchemaFiles, password); err != nil {
-			return nil, fmt.Errorf("branch %s has stale changes from a previous apply — delete the branch and retry without --branch: %w", branchName, err)
-		}
-	}
-
-	// Apply DDL and VSchema changes to all keyspaces
-	emitEvent(engine.ApplyEvent{
-		Message:  "Applying changes to branch",
-		Metadata: map[string]string{"branch": branchName},
-		NewState: state.Apply.ApplyingBranchChanges,
-	})
-	if err := e.applyChangesToBranch(ctx, req.Changes, req.SchemaFiles, password, client, org, req.Database, branchName, emitEvent); err != nil {
-		return nil, fmt.Errorf("apply changes to branch: %w", err)
-	}
-	ddlCount := 0
-	for _, sc := range req.Changes {
-		ddlCount += len(sc.TableChanges)
-	}
-	emitEvent(engine.ApplyEvent{
-		Message:  fmt.Sprintf("Applied %d DDL changes to branch %s", ddlCount, branchName),
-		Metadata: map[string]string{"branch": branchName},
-		NewState: state.Apply.ValidatingBranch,
-	})
-
-	// Verify the branch now matches the desired schema. If DDL application
-	// was partial or the branch had stale state, some tables may still differ.
-	// Catch this before creating the deploy request to prevent deploying
-	// unexpected changes (e.g., DROP COLUMN when ADD COLUMN was intended).
-	//
-	// DDL is fetched via MySQL (real-time), but VSchema is fetched via the
-	// PlanetScale API which may return stale data after UpdateKeyspaceVSchema.
-	// Retry up to 30s to allow the API to converge.
-	keyspaces := sortedKeyspaces(req.SchemaFiles)
-	if err := e.verifyBranchMatchesDesiredWithRetry(ctx, client, org, req.Database, branchName, keyspaces, req.SchemaFiles, password); err != nil {
-		return nil, fmt.Errorf("branch validation failed after DDL apply: %w", err)
-	}
-	emitEvent(engine.ApplyEvent{
-		Message:  "Branch schema validated — matches desired state",
-		Metadata: map[string]string{"branch": branchName},
-		NewState: state.Apply.CreatingDeployRequest,
-	})
-
-	// Capture existing migration_contexts before deploy so we can discover the new one
-	existingContexts := e.captureExistingContexts(ctx, client, req.Database, req.Credentials)
-
-	// Check defer options
-	deferCutover := req.Options["defer_cutover"] == "true"
-	deferDeploy := req.Options["defer_deploy"] == "true"
-
-	// Create deploy request and wait for it to be ready.
-	// The server computes the schema diff asynchronously — poll until the deploy
-	// request transitions from "pending" to "ready" (or "no_changes"/"error").
-	drStart := time.Now()
-	autoDeleteBranch := existingBranch == "" // don't delete reused branches
-	dr, err := e.createDeployRequest(ctx, client, org, req.Database, branchName, main, !deferCutover, autoDeleteBranch)
-	if err != nil {
-		return nil, fmt.Errorf("create deploy request: %w", err)
-	}
-	emitEvent(engine.ApplyEvent{
-		Message: fmt.Sprintf("Deploy request #%d created, validating...", dr.Number),
-		Metadata: map[string]string{
-			"deploy_request_id":  fmt.Sprintf("%d", dr.Number),
-			"deploy_request_url": dr.HtmlURL,
-			"branch":             branchName,
-		},
-		NewState: state.Apply.ValidatingDeployRequest,
-	})
-	persistState(&psMetadata{
-		BranchName:       branchName,
-		DeployRequestID:  dr.Number,
-		DeployRequestURL: dr.HtmlURL,
-	})
-	for dr.DeploymentState == deployState.Pending {
-		select {
-		case <-ctx.Done():
-			return nil, fmt.Errorf("context cancelled waiting for deploy request: %w", ctx.Err())
-		case <-time.After(500 * time.Millisecond):
-		}
-		dr, err = e.getDeployRequest(ctx, client, org, req.Database, dr.Number)
-		if err != nil {
-			return nil, fmt.Errorf("poll deploy request %d: %w", dr.Number, err)
-		}
-	}
-	if dr.DeploymentState == deployState.Error {
-		errMsg := formatDeployRequestError(dr)
-		emitEvent(engine.ApplyEvent{
-			Message:  errMsg,
-			Metadata: map[string]string{"deploy_request_id": fmt.Sprintf("%d", dr.Number)},
-		})
-		return nil, fmt.Errorf("%s", errMsg)
-	}
-	if dr.DeploymentState == deployState.NoChanges {
-		emitEvent(engine.ApplyEvent{
-			Message:  fmt.Sprintf("Deploy request #%d: no changes detected", dr.Number),
-			Metadata: map[string]string{"deploy_request_id": fmt.Sprintf("%d", dr.Number)},
-		})
-		return &engine.ApplyResult{Message: "no changes detected"}, nil
-	}
-
-	// Determine instant DDL eligibility. Prefer instant when PlanetScale reports
-	// it as eligible — instant DDL modifies metadata only (no row copy), so it
-	// completes immediately and has no revert window regardless of skip_revert.
-	// Instant DDL is orthogonal to defer flags — the mechanism (instant vs row copy)
-	// is independent of when the deploy executes.
-	instantEligible := dr.Deployment != nil && dr.Deployment.InstantDDLEligible
-	useInstant := instantEligible
-
-	// Log the raw deploy request fields for debugging instant DDL detection.
-	if dr.Deployment != nil {
-		e.logger.Info("deploy request deployment info",
-			"database", req.Database,
-			"deploy_request", dr.Number,
-			"instant_ddl_eligible", dr.Deployment.InstantDDLEligible,
-			"deployment_state", dr.Deployment.State,
-		)
-	} else {
-		e.logger.Warn("deploy request has nil deployment",
-			"database", req.Database,
-			"deploy_request", dr.Number,
-			"deploy_state", dr.DeploymentState,
-		)
-	}
-	e.logger.Info("instant DDL decision",
-		"database", req.Database,
-		"deploy_request", dr.Number,
-		"has_deployment", dr.Deployment != nil,
-		"instant_eligible", instantEligible,
-		"use_instant", useInstant,
-		"defer_cutover", deferCutover,
-		"defer_deploy", deferDeploy,
-		"deploy_state", dr.DeploymentState,
-	)
-
-	// Log when --defer-cutover has no effect for instant DDL
-	if deferCutover && useInstant {
-		e.logger.Info("--defer-cutover has no effect for instant DDL",
-			"database", req.Database,
-			"deploy_request", dr.Number,
-		)
-		emitEvent(engine.ApplyEvent{
-			Message: "Note: --defer-cutover has no effect for instant DDL",
-		})
-	}
-
-	drElapsed := time.Since(drStart).Round(time.Second)
-	readyMsg := fmt.Sprintf("Deploy request #%d ready (%s)", dr.Number, drElapsed)
-	if useInstant {
-		readyMsg += " — instant DDL eligible"
-	}
-	emitEvent(engine.ApplyEvent{
-		Message: readyMsg,
-		Metadata: map[string]string{
-			"deploy_request_id":  fmt.Sprintf("%d", dr.Number),
-			"deploy_request_url": dr.HtmlURL,
-			"instant_ddl":        fmt.Sprintf("%t", useInstant),
-		},
-	})
-
-	// Deferred deploy: don't call DeployDeployRequest yet. The user will review
-	// the deploy request diff on PlanetScale and trigger via `schemabot cutover`.
-	if deferDeploy {
-		e.logger.Info("deferring deploy — user must trigger via cutover",
-			"database", req.Database,
-			"deploy_request", dr.Number,
-			"instant_eligible", useInstant,
-		)
-		meta, encErr := encodePSMetadata(&psMetadata{
-			BranchName:       branchName,
-			DeployRequestID:  dr.Number,
-			DeployRequestURL: dr.HtmlURL,
-			IsInstant:        useInstant,
-			DeferredDeploy:   true,
-		})
-		if encErr != nil {
-			return nil, fmt.Errorf("encode metadata for deferred deploy request #%d: %w", dr.Number, encErr)
-		}
-		suffix := ""
-		if useInstant {
-			suffix = " (instant DDL)"
-		}
-		return &engine.ApplyResult{
-			Accepted: true,
-			Message:  fmt.Sprintf("Deploy request #%d ready%s — waiting for deploy", dr.Number, suffix),
-			ResumeState: &engine.ResumeState{
-				MigrationContext: migCtx,
-				Metadata:         meta,
-			},
-		}, nil
-	}
-
-	// Deploy (starts the schema change). PlanetScale may still be validating
-	// the deploy request even after reporting "ready", so retry on transient
-	// validation errors.
-	drNumber := dr.Number
-	var deployErr error
-	for attempt := range maxRetries {
-		if attempt > 0 {
-			delay := retryDelay(attempt, deployErr)
-			e.logger.Warn("retrying deploy request", "deploy_request", drNumber, "attempt", attempt+1, "delay", delay.Round(time.Millisecond), "error", deployErr)
-			select {
-			case <-ctx.Done():
-				return nil, ctx.Err()
-			case <-time.After(delay):
-			}
-		}
-		dr, deployErr = client.DeployDeployRequest(ctx, &ps.PerformDeployRequest{
-			Organization: org,
-			Database:     req.Database,
-			Number:       drNumber,
-			InstantDDL:   useInstant,
-		})
-		if deployErr == nil {
-			break
-		}
-		if strings.Contains(deployErr.Error(), "approved") {
-			return nil, fmt.Errorf("deploy request #%d could not be deployed: PlanetScale deploy request approvals are not supported — disable 'Require administrator approval for deploy requests' in the PlanetScale database settings", drNumber)
-		}
-		if !isRetryablePSError(deployErr) {
-			return nil, fmt.Errorf("deploy deploy request #%d: %w", drNumber, deployErr)
-		}
-	}
-	if deployErr != nil {
-		return nil, fmt.Errorf("deploy deploy request #%d (after %d attempts): %w", drNumber, maxRetries, deployErr)
-	}
-
-	emitEvent(engine.ApplyEvent{
-		Message: fmt.Sprintf("Deploy request #%d deployed", dr.Number),
-		Metadata: map[string]string{
-			"deploy_request_id": fmt.Sprintf("%d", dr.Number),
-			"instant_ddl":       fmt.Sprintf("%t", useInstant),
-		},
-	})
-
-	// Discover migration_context by diffing current SHOW VITESS_MIGRATIONS against
-	// the pre-deploy baseline. Retries because Vitess may not have created migrations
-	// immediately after the deploy request is submitted.
-	var migrationContext string
-	for attempt := range 10 {
-		migrationContext = e.discoverMigrationContext(ctx, client, req.Database, req.Credentials, existingContexts)
-		if migrationContext != "" {
-			break
-		}
-		if attempt < 9 {
-			time.Sleep(500 * time.Millisecond)
-		}
-	}
-
-	meta, err := encodePSMetadata(&psMetadata{
-		BranchName:       branchName,
-		DeployRequestID:  dr.Number,
-		DeployRequestURL: dr.HtmlURL,
-		IsInstant:        useInstant,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("encode metadata for deploy request #%d: %w", dr.Number, err)
-	}
-
-	return &engine.ApplyResult{
-		Accepted: true,
-		Message:  fmt.Sprintf("Deploy request #%d created", dr.Number),
-		ResumeState: &engine.ResumeState{
-			MigrationContext: migrationContext,
-			Metadata:         meta,
-		},
-	}, nil
-}
-
-// applyChangesToBranch applies VSchema and DDL changes to all keyspaces.
-// VSchema updates are applied sequentially (PlanetScale rejects concurrent
-// VSchema writes during schema snapshots). DDL is applied in parallel after
-// all VSchema changes are committed.
-func (e *Engine) applyChangesToBranch(ctx context.Context, changes []engine.SchemaChange, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword, client psclient.PSClient, org, database, branchName string, emitEvent func(engine.ApplyEvent)) error {
-	if len(changes) == 0 {
-		e.logger.Debug("no changes to apply to branch", "branch", branchName)
-		return nil
-	}
-
-	total := len(changes)
-	var applied atomic.Int32
-
-	// Serialize event callbacks — OnEvent mutates shared apply state.
-	var eventMu sync.Mutex
-	safeEmit := func(event engine.ApplyEvent) {
-		eventMu.Lock()
-		defer eventMu.Unlock()
-		emitEvent(event)
-	}
-
-	safeEmit(engine.ApplyEvent{
-		Message:  fmt.Sprintf("Applying changes to %d keyspaces on branch %s", total, branchName),
-		Metadata: map[string]string{"branch": branchName},
-		NewState: state.Apply.ApplyingBranchChanges,
-	})
-
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(maxConcurrentKeyspaces)
-	for _, sc := range changes {
-		g.Go(func() error {
-			if err := e.applyKeyspaceChanges(gCtx, sc, schemaFiles, password, client, org, database, branchName); err != nil {
-				return err
-			}
-			n := int(applied.Add(1))
-			safeEmit(engine.ApplyEvent{
-				Message:  fmt.Sprintf("Applied keyspace %s (%d/%d)", sc.Namespace, n, total),
-				Metadata: map[string]string{"keyspace": sc.Namespace},
-			})
-			return nil
-		})
-	}
-	return g.Wait()
-}
-
-// applyKeyspaceChanges applies VSchema and DDL for a single keyspace with retries.
-// Uses longer backoff when PlanetScale reports a schema snapshot is in progress.
-func (e *Engine) applyKeyspaceChanges(ctx context.Context, sc engine.SchemaChange, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword, client psclient.PSClient, org, database, branchName string) error {
-	start := time.Now()
-	e.logger.Info(fmt.Sprintf("applying changes to keyspace %s on branch %s", sc.Namespace, branchName),
-		"keyspace", sc.Namespace,
-		"ddl_count", len(sc.TableChanges),
-		"has_vschema", sc.Metadata["vschema_changed"] == "true",
-		"branch", branchName,
-	)
-
-	maxAttempts := maxRetries
-	var lastErr error
-	for attempt := 0; attempt < maxAttempts; attempt++ {
-		if attempt > 0 {
-			delay := retryDelay(attempt, lastErr)
-			e.logger.Warn("retrying keyspace apply", "keyspace", sc.Namespace, "attempt", attempt+1, "delay", delay.Round(time.Millisecond), "error", lastErr)
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case <-time.After(delay):
-			}
-		}
-
-		if err := e.applyKeyspaceChangesOnce(ctx, sc, schemaFiles, password, client, org, database, branchName); err != nil {
-			lastErr = err
-			e.logger.Error(fmt.Sprintf("keyspace %s apply attempt %d failed", sc.Namespace, attempt+1), "keyspace", sc.Namespace, "attempt", attempt+1, "error", err)
-			if !isRetryableEngineError(err) {
-				return engine.NewPermanentError("apply keyspace %s: %w", sc.Namespace, err)
-			}
-			if isSnapshotInProgress(err) && maxAttempts == maxRetries {
-				maxAttempts = maxSnapshotRetries
-				e.logger.Info("schema snapshot in progress, extending retries",
-					"keyspace", sc.Namespace, "max_attempts", maxAttempts)
-			}
-			continue
-		}
-		e.logger.Info(fmt.Sprintf("keyspace %s changes applied (%s)", sc.Namespace, time.Since(start).Round(time.Second)), "keyspace", sc.Namespace, "elapsed", time.Since(start).Round(time.Second))
-		return nil
-	}
-	finalErr := fmt.Errorf("apply keyspace %s (after %d attempts): %w", sc.Namespace, maxAttempts, lastErr)
-	return finalErr
-}
-
 // isRetryableEngineError returns true if the error is a transient condition
 // that may succeed on a future attempt.
 func isRetryableEngineError(err error) bool {
@@ -1376,1672 +702,6 @@ func retryDelay(attempt int, lastErr error) time.Duration {
 	return base + time.Duration(rand.IntN(2000))*time.Millisecond
 }
 
-// applyKeyspaceChangesOnce applies VSchema and DDL for a single keyspace in one attempt.
-func (e *Engine) applyKeyspaceChangesOnce(ctx context.Context, sc engine.SchemaChange, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword, client psclient.PSClient, org, database, branchName string) error {
-	// Apply VSchema first — vtgate needs VSchema to route DDL correctly
-	if vschemaContent := getVSchemaContent(sc, schemaFiles); vschemaContent != "" {
-		if err := e.updateBranchVSchema(ctx, client, org, database, branchName, sc.Namespace, vschemaContent); err != nil {
-			return fmt.Errorf("update vschema for %s: %w", sc.Namespace, err)
-		}
-		e.logger.Info(fmt.Sprintf("applied vschema for %s on branch %s", sc.Namespace, branchName), "keyspace", sc.Namespace, "branch", branchName)
-	}
-
-	if len(sc.TableChanges) == 0 {
-		e.logger.Debug("no DDL for keyspace, vschema-only", "keyspace", sc.Namespace, "branch", branchName)
-		return nil
-	}
-
-	// Build DSN targeting this specific keyspace.
-	// TLS is configured automatically when RegisterMTLS has been called.
-	mysqlCfg := mysql.NewConfig()
-	mysqlCfg.User = password.Username
-	mysqlCfg.Passwd = password.PlainText
-	mysqlCfg.Net = "tcp"
-	mysqlCfg.Addr = password.Hostname
-	mysqlCfg.InterpolateParams = true
-	if mtlsRegistered.Load() {
-		mysqlCfg.TLSConfig = mtlsConfigName
-	}
-	dsn := mysqlCfg.FormatDSN()
-
-	db, err := sql.Open("mysql", dsn)
-	if err != nil {
-		return fmt.Errorf("open branch connection for %s: %w", sc.Namespace, err)
-	}
-	defer utils.CloseAndLog(db)
-
-	if err := db.PingContext(ctx); err != nil {
-		return fmt.Errorf("ping branch for %s: %w", sc.Namespace, err)
-	}
-
-	// USE the keyspace — vtgate branch proxy routes based on the database name.
-	if _, err := db.ExecContext(ctx, "USE `"+sc.Namespace+"`"); err != nil {
-		return fmt.Errorf("use keyspace %s: %w", sc.Namespace, err)
-	}
-
-	for _, tc := range sc.TableChanges {
-		e.logger.Info(fmt.Sprintf("applying DDL to %s.%s on branch", sc.Namespace, tc.Table),
-			"keyspace", sc.Namespace,
-			"table", tc.Table,
-			"operation", tc.Operation,
-			"ddl", tc.DDL,
-		)
-		if _, err := db.ExecContext(ctx, tc.DDL); err != nil {
-			return fmt.Errorf("execute DDL on %s.%s: %w\nstatement: %s", sc.Namespace, tc.Table, err, tc.DDL)
-		}
-	}
-	return nil
-}
-
-// getVSchemaContent extracts the vschema.json content for a keyspace from schema files.
-// Returns empty string if no VSchema change is needed.
-func getVSchemaContent(sc engine.SchemaChange, schemaFiles schema.SchemaFiles) string {
-	if sc.Metadata["vschema_changed"] != "true" {
-		return ""
-	}
-	if ns, ok := schemaFiles[sc.Namespace]; ok && ns != nil {
-		if content, ok := ns.Files["vschema.json"]; ok {
-			return content
-		}
-	}
-	return ""
-}
-
-// updateBranchVSchema updates the VSchema for a keyspace on a branch
-// using the PlanetScale SDK's UpdateKeyspaceVSchema endpoint.
-func (e *Engine) updateBranchVSchema(ctx context.Context, client psclient.PSClient, org, database, branch, keyspace, vschemaJSON string) error {
-	e.logger.Info(fmt.Sprintf("updating VSchema for %s on branch %s", keyspace, branch),
-		"branch", branch,
-		"keyspace", keyspace,
-	)
-	_, err := client.UpdateKeyspaceVSchema(ctx, &ps.UpdateKeyspaceVSchemaRequest{
-		Organization: org,
-		Database:     database,
-		Branch:       branch,
-		Keyspace:     keyspace,
-		VSchema:      vschemaJSON,
-	})
-	if err != nil {
-		return fmt.Errorf("update vschema for keyspace %s on branch %s: %w", keyspace, branch, err)
-	}
-	return nil
-}
-
-// diffBranchForResume fetches the working branch's current schema and diffs it
-// against the desired schema to find DDL that wasn't applied before the crash.
-func (e *Engine) diffBranchForResume(ctx context.Context, client psclient.PSClient, org, database, branch string, schemaFiles schema.SchemaFiles) ([]engine.SchemaChange, error) {
-	currentSchema, err := e.fetchDatabaseSchema(ctx, client, org, database, branch, sortedKeyspaces(schemaFiles))
-	if err != nil {
-		return nil, fmt.Errorf("fetch branch schema: %w", err)
-	}
-
-	var changes []engine.SchemaChange
-	for _, keyspace := range sortedKeyspaces(schemaFiles) {
-		ns := schemaFiles[keyspace]
-
-		// Build current table schemas from branch
-		var currentTableSchemas []table.TableSchema
-		if tables, ok := currentSchema[keyspace]; ok {
-			currentTableSchemas = append(currentTableSchemas, tables...)
-		}
-
-		// Build desired table schemas from files
-		desiredTableSchemas, err := parseDesiredSchemas(keyspace, ns)
-		if err != nil {
-			return nil, err
-		}
-
-		// Diff: what DDL is needed to bring branch from current to desired?
-		plan, err := lint.PlanChanges(currentTableSchemas, desiredTableSchemas, nil, e.linter.SpiritConfig())
-		if err != nil {
-			return nil, fmt.Errorf("diff keyspace %s for resume: %w", keyspace, err)
-		}
-		if !plan.HasChanges() {
-			continue
-		}
-
-		sc := engine.SchemaChange{
-			Namespace: keyspace,
-			Metadata:  make(map[string]string),
-		}
-		for _, pc := range plan.Changes {
-			stmtType, _, classifyErr := ddl.ClassifyStatement(pc.Statement)
-			if classifyErr != nil {
-				return nil, fmt.Errorf("classify statement in keyspace %s: %w", keyspace, classifyErr)
-			}
-			sc.TableChanges = append(sc.TableChanges, engine.TableChange{
-				Table:     pc.TableName,
-				Operation: stmtType,
-				DDL:       pc.Statement,
-			})
-		}
-		changes = append(changes, sc)
-	}
-	return changes, nil
-}
-
-// resumeApply resumes a schema change after restart.
-// Handles two crash scenarios:
-//   - Branch exists, no deploy request: diff branch against desired schema, apply remaining DDL, create deploy request
-//   - Branch exists, deploy request exists: just return current state for Progress polling
-func (e *Engine) resumeApply(ctx context.Context, client psclient.PSClient, org string, req *engine.ApplyRequest) (*engine.ApplyResult, error) {
-	meta, err := decodePSMetadata(req.ResumeState.Metadata)
-	if err != nil {
-		return nil, fmt.Errorf("decode resume state: %w", err)
-	}
-
-	emitEvent := func(event engine.ApplyEvent) {
-		attrs := []any{"database", req.Database}
-		for k, v := range event.Metadata {
-			attrs = append(attrs, k, v)
-		}
-		e.logger.Info(event.Message, attrs...)
-		if req.OnEvent != nil {
-			req.OnEvent(event)
-		}
-	}
-
-	e.logger.Info("resuming apply",
-		"branch", meta.BranchName,
-		"deploy_request", meta.DeployRequestID,
-	)
-
-	// If we have a deploy request ID, check its current state.
-	if meta.DeployRequestID != 0 {
-		dr, err := e.getDeployRequest(ctx, client, org, req.Database, meta.DeployRequestID)
-		if err != nil {
-			// Deploy request may have been cleaned up — start fresh.
-			e.logger.Warn("deploy request not found on resume, starting fresh",
-				"deploy_request", meta.DeployRequestID, "error", err)
-			req.ResumeState = nil
-			return e.Apply(ctx, req)
-		}
-
-		// If the deploy request failed, start fresh with a new branch rather
-		// than resuming a broken deploy.
-		if dr.DeploymentState == deployState.Error || dr.DeploymentState == deployState.CompleteError {
-			e.logger.Warn("deploy request in error state on resume, starting fresh",
-				"deploy_request", meta.DeployRequestID, "state", dr.DeploymentState)
-			req.ResumeState = nil
-			return e.Apply(ctx, req)
-		}
-
-		meta.DeployRequestURL = dr.HtmlURL
-		updatedMeta, err := encodePSMetadata(meta)
-		if err != nil {
-			return nil, fmt.Errorf("encode metadata for deploy request #%d: %w", meta.DeployRequestID, err)
-		}
-		return &engine.ApplyResult{
-			Accepted: true,
-			Message:  fmt.Sprintf("Resumed deploy request #%d (state: %s)", dr.Number, dr.DeploymentState),
-			ResumeState: &engine.ResumeState{
-				MigrationContext: req.ResumeState.MigrationContext,
-				Metadata:         updatedMeta,
-			},
-		}, nil
-	}
-
-	// No deploy request yet — worker crashed after branch creation but before
-	// the deploy request was created. Diff the branch against desired schema
-	// to find DDL that wasn't applied before the crash, then apply only the
-	// missing changes.
-	e.logger.Info("resuming from branch (no deploy request yet)", "branch", meta.BranchName)
-
-	// Check if the branch still exists — it may have been deleted by TTL
-	// between the crash and recovery. If so, start fresh.
-	if err := e.waitForBranchReady(ctx, client, org, req.Database, meta.BranchName); err != nil {
-		e.logger.Warn("branch no longer available on resume, starting fresh", "branch", meta.BranchName, "error", err)
-		req.ResumeState = nil
-		return e.Apply(ctx, req)
-	}
-
-	// Diff branch's current state against desired to find un-applied DDL
-	remainingChanges, err := e.diffBranchForResume(ctx, client, org, req.Database, meta.BranchName, req.SchemaFiles)
-	if err != nil {
-		return nil, fmt.Errorf("diff branch for resume: %w", err)
-	}
-
-	if len(remainingChanges) > 0 {
-		e.logger.Info("applying remaining DDL on resume", "branch", meta.BranchName, "keyspaces", len(remainingChanges))
-		resumePwCtx, resumePwCancel := context.WithTimeout(ctx, 10*time.Second)
-		defer resumePwCancel()
-
-		password, err := client.CreateBranchPassword(resumePwCtx, &ps.DatabaseBranchPasswordRequest{
-			Organization: org, Database: req.Database, Branch: meta.BranchName, Role: "admin", TTL: 3600,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("create branch password on resume: %w", err)
-		}
-		if err := e.applyChangesToBranch(ctx, remainingChanges, req.SchemaFiles, password, client, org, req.Database, meta.BranchName, emitEvent); err != nil {
-			return nil, fmt.Errorf("apply remaining DDL on resume: %w", err)
-		}
-	} else {
-		e.logger.Info("all DDL already applied on branch", "branch", meta.BranchName)
-	}
-
-	// VSchema may not have been applied before the crash — re-apply
-	// (VSchema updates are idempotent, they overwrite the entire VSchema)
-	for _, sc := range req.Changes {
-		if vschemaContent := getVSchemaContent(sc, req.SchemaFiles); vschemaContent != "" {
-			if err := e.updateBranchVSchema(ctx, client, org, req.Database, meta.BranchName, sc.Namespace, vschemaContent); err != nil {
-				return nil, fmt.Errorf("update vschema for %s on resume: %w", sc.Namespace, err)
-			}
-		}
-	}
-
-	// Create deploy request
-	main := mainBranch(req.Credentials)
-	deferCutover := req.Options["defer_cutover"] == "true"
-	deferDeploy := req.Options["defer_deploy"] == "true"
-
-	dr, err := e.createDeployRequest(ctx, client, org, req.Database, meta.BranchName, main, !deferCutover, true)
-	if err != nil {
-		return nil, fmt.Errorf("create deploy request on resume: %w", err)
-	}
-	for dr.DeploymentState == deployState.Pending {
-		time.Sleep(500 * time.Millisecond)
-		dr, err = e.getDeployRequest(ctx, client, org, req.Database, dr.Number)
-		if err != nil {
-			return nil, fmt.Errorf("poll deploy request %d on resume: %w", dr.Number, err)
-		}
-	}
-	if dr.DeploymentState == deployState.Error {
-		return nil, fmt.Errorf("deploy request #%d failed on resume (state: %s)", dr.Number, dr.DeploymentState)
-	}
-	if dr.DeploymentState == deployState.NoChanges {
-		return &engine.ApplyResult{Message: "no changes detected on resume"}, nil
-	}
-
-	// Deploy — prefer instant when eligible (no row copy, no revert window needed).
-	instantEligible := dr.Deployment != nil && dr.Deployment.InstantDDLEligible
-	useInstant := instantEligible
-
-	meta.DeployRequestID = dr.Number
-	meta.DeployRequestURL = dr.HtmlURL
-	meta.IsInstant = useInstant
-
-	// Deferred deploy on resume: don't start the deploy yet.
-	if deferDeploy {
-		meta.DeferredDeploy = true
-		persistMeta, encErr := encodePSMetadata(meta)
-		if encErr != nil {
-			return nil, fmt.Errorf("encode metadata for deferred deploy on resume: %w", encErr)
-		}
-		if req.OnStateChange != nil {
-			req.OnStateChange(&engine.ResumeState{
-				MigrationContext: req.ResumeState.MigrationContext,
-				Metadata:         persistMeta,
-			})
-		}
-		suffix := ""
-		if useInstant {
-			suffix = " (instant DDL)"
-		}
-		return &engine.ApplyResult{
-			Accepted: true,
-			Message:  fmt.Sprintf("Deploy request #%d ready%s — waiting for deploy", dr.Number, suffix),
-			ResumeState: &engine.ResumeState{
-				MigrationContext: req.ResumeState.MigrationContext,
-				Metadata:         persistMeta,
-			},
-		}, nil
-	}
-
-	persistMeta, err := encodePSMetadata(meta)
-	if err != nil {
-		return nil, fmt.Errorf("encode metadata on resume: %w", err)
-	}
-	if req.OnStateChange != nil {
-		req.OnStateChange(&engine.ResumeState{
-			MigrationContext: req.ResumeState.MigrationContext,
-			Metadata:         persistMeta,
-		})
-	}
-
-	dr, err = client.DeployDeployRequest(ctx, &ps.PerformDeployRequest{
-		Organization: org, Database: req.Database, Number: dr.Number, InstantDDL: useInstant,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("deploy on resume: %w", err)
-	}
-
-	e.logger.Info("resumed and deployed", "number", dr.Number, "branch", meta.BranchName)
-	return &engine.ApplyResult{
-		Accepted: true,
-		Message:  fmt.Sprintf("Resumed and deployed request #%d", dr.Number),
-		ResumeState: &engine.ResumeState{
-			MigrationContext: req.ResumeState.MigrationContext,
-			Metadata:         persistMeta,
-		},
-	}, nil
-}
-
-// --- Progress ---
-
-// Progress polls deploy request status from PlanetScale's API and optionally queries
-// SHOW VITESS_MIGRATIONS for per-table, per-shard row counts and ETA.
-func (e *Engine) Progress(ctx context.Context, req *engine.ProgressRequest) (*engine.ProgressResult, error) {
-	if req.ResumeState == nil || req.ResumeState.Metadata == "" {
-		return &engine.ProgressResult{
-			State:   engine.StatePending,
-			Message: "No active schema change",
-		}, nil
-	}
-
-	meta, err := decodePSMetadata(req.ResumeState.Metadata)
-	if err != nil {
-		return nil, fmt.Errorf("decode resume state: %w", err)
-	}
-
-	if meta.DeployRequestID == 0 {
-		return &engine.ProgressResult{
-			State:   engine.StatePending,
-			Message: fmt.Sprintf("Setting up branch %s", meta.BranchName),
-		}, nil
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	dr, err := e.getDeployRequest(ctx, client, credOrg(req.Credentials), req.Database, meta.DeployRequestID)
-	if err != nil {
-		// Deploy request not found is permanent (e.g., server restarted
-		// with fresh LocalScale state but stale apply record).
-		var psErr *ps.Error
-		if errors.As(err, &psErr) && psErr.Code == ps.ErrNotFound {
-			return nil, engine.NewPermanentError("deploy request #%d not found: %w", meta.DeployRequestID, err)
-		}
-		return nil, fmt.Errorf("get deploy request: %w", err)
-	}
-
-	engineState := deployStateToEngineState(dr.DeploymentState)
-
-	// Deferred deploy: the deploy request is ready but hasn't been triggered yet.
-	if meta.DeferredDeploy && dr.DeploymentState == deployState.Ready {
-		engineState = engine.StateWaitingForDeploy
-	}
-
-	// Update instant DDL flag from deploy request if not already set.
-	if !meta.IsInstant && dr.Deployment != nil && dr.Deployment.InstantDDLEligible {
-		meta.IsInstant = true
-	}
-
-	// Update metadata with DeployedAt when available (used by tern layer for
-	// revert window timeout calculation).
-	if dr.DeployedAt != nil && meta.DeployedAt == nil {
-		meta.DeployedAt = dr.DeployedAt
-		if encoded, encErr := encodePSMetadata(meta); encErr == nil {
-			req.ResumeState = &engine.ResumeState{
-				MigrationContext: req.ResumeState.MigrationContext,
-				Metadata:         encoded,
-			}
-		}
-	}
-
-	e.logger.Debug("progress poll",
-		"database", req.Database,
-		"deploy_request", meta.DeployRequestID,
-		"deploy_state", dr.DeploymentState,
-		"engine_state", engineState,
-		"is_instant", meta.IsInstant,
-		"has_migration_context", req.ResumeState != nil && req.ResumeState.MigrationContext != "",
-		"has_vtgate_dsn", req.Credentials.DSN != "",
-	)
-
-	result := &engine.ProgressResult{
-		State:       engineState,
-		Message:     deployStateToMessage(dr.DeploymentState),
-		ResumeState: req.ResumeState,
-	}
-
-	// Enrich with per-shard progress from SHOW VITESS_MIGRATIONS.
-	// Requires a vtgate DSN (Credentials.DSN) and a migration context
-	// (from VitessApplyData) to query per-shard state.
-	hasMigrationContext := req.Credentials.DSN != "" &&
-		req.ResumeState != nil && req.ResumeState.MigrationContext != ""
-	if hasMigrationContext {
-		tables, overallProgress := e.queryVitessMigrations(ctx, client, req.Database, req.Credentials, req.ResumeState.MigrationContext)
-		e.logger.Debug("vitess migrations queried",
-			"database", req.Database,
-			"table_count", len(tables),
-			"overall_progress", overallProgress,
-		)
-		if len(tables) > 0 {
-			result.Tables = tables
-			if overallProgress > 0 {
-				result.Progress = overallProgress
-			}
-		}
-	} else {
-		e.logger.Debug("skipping per-shard progress",
-			"database", req.Database,
-			"has_vtgate_dsn", req.Credentials.DSN != "",
-			"has_migration_context", req.ResumeState != nil && req.ResumeState.MigrationContext != "",
-		)
-	}
-
-	// Propagate instant DDL flag to all tables. Instant DDL may complete
-	// before migration context discovery, so we use the flag from deploy
-	// metadata as the authoritative source.
-	if meta.IsInstant {
-		e.logger.Debug("marking tables as instant DDL",
-			"database", req.Database,
-			"table_count", len(result.Tables),
-		)
-		for i := range result.Tables {
-			result.Tables[i].IsInstant = true
-		}
-	}
-
-	return result, nil
-}
-
-// --- Control operations ---
-
-// Stop cancels the deploy request. This is permanent.
-func (e *Engine) Stop(ctx context.Context, req *engine.ControlRequest) (*engine.ControlResult, error) {
-	meta, err := e.controlMeta(req)
-	if err != nil {
-		return nil, fmt.Errorf("decode control metadata: %w", err)
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	_, err = client.CancelDeployRequest(ctx, &ps.CancelDeployRequestRequest{
-		Organization: credOrg(req.Credentials),
-		Database:     req.Database,
-		Number:       meta.DeployRequestID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("cancel deploy request #%d (may have been deleted): %w", meta.DeployRequestID, err)
-	}
-
-	return &engine.ControlResult{
-		Accepted:    true,
-		Message:     "Deploy request cancelled",
-		ResumeState: req.ResumeState,
-	}, nil
-}
-
-// Start starts a deferred deploy request. Cancelled deploy requests cannot be restarted.
-func (e *Engine) Start(ctx context.Context, req *engine.ControlRequest) (*engine.ControlResult, error) {
-	meta, err := e.controlMeta(req)
-	if err != nil {
-		return nil, fmt.Errorf("decode control metadata: %w", err)
-	}
-
-	if !meta.DeferredDeploy {
-		return nil, fmt.Errorf("start not supported for planetscale engine: cancelled deploy requests cannot be restarted")
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	e.logger.Info("starting deferred deploy",
-		"deploy_request", meta.DeployRequestID,
-		"instant_ddl", meta.IsInstant,
-	)
-	dr, deployErr := client.DeployDeployRequest(ctx, &ps.PerformDeployRequest{
-		Organization: credOrg(req.Credentials),
-		Database:     req.Database,
-		Number:       meta.DeployRequestID,
-		InstantDDL:   meta.IsInstant,
-	})
-	if deployErr != nil {
-		return nil, fmt.Errorf("deploy deploy request #%d: %w", meta.DeployRequestID, deployErr)
-	}
-	return &engine.ControlResult{
-		Accepted:    true,
-		Message:     fmt.Sprintf("Deploy initiated for deploy request #%d", dr.Number),
-		ResumeState: req.ResumeState,
-	}, nil
-}
-
-// Cutover triggers the final schema swap via ApplyDeployRequest.
-func (e *Engine) Cutover(ctx context.Context, req *engine.ControlRequest) (*engine.ControlResult, error) {
-	meta, err := e.controlMeta(req)
-	if err != nil {
-		return nil, fmt.Errorf("decode control metadata: %w", err)
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	dr, err := client.ApplyDeployRequest(ctx, &ps.ApplyDeployRequestRequest{
-		Organization: credOrg(req.Credentials),
-		Database:     req.Database,
-		Number:       meta.DeployRequestID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("cutover deploy request #%d (may have been deleted): %w", meta.DeployRequestID, err)
-	}
-
-	return &engine.ControlResult{
-		Accepted:    true,
-		Message:     fmt.Sprintf("Cutover initiated for deploy request #%d", dr.Number),
-		ResumeState: req.ResumeState,
-	}, nil
-}
-
-// Revert rolls back a completed schema change during the revert window.
-func (e *Engine) Revert(ctx context.Context, req *engine.ControlRequest) (*engine.ControlResult, error) {
-	meta, err := e.controlMeta(req)
-	if err != nil {
-		return nil, fmt.Errorf("decode control metadata: %w", err)
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	dr, err := client.RevertDeployRequest(ctx, &ps.RevertDeployRequestRequest{
-		Organization: credOrg(req.Credentials),
-		Database:     req.Database,
-		Number:       meta.DeployRequestID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("revert deploy request #%d (may have been deleted): %w", meta.DeployRequestID, err)
-	}
-
-	return &engine.ControlResult{
-		Accepted:    true,
-		Message:     fmt.Sprintf("Revert initiated for deploy request #%d", dr.Number),
-		ResumeState: req.ResumeState,
-	}, nil
-}
-
-// SkipRevert closes the revert window, making the schema change permanent.
-func (e *Engine) SkipRevert(ctx context.Context, req *engine.ControlRequest) (*engine.ControlResult, error) {
-	meta, err := e.controlMeta(req)
-	if err != nil {
-		return nil, fmt.Errorf("decode control metadata: %w", err)
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	dr, err := client.SkipRevertDeployRequest(ctx, &ps.SkipRevertDeployRequestRequest{
-		Organization: credOrg(req.Credentials),
-		Database:     req.Database,
-		Number:       meta.DeployRequestID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("skip revert for deploy request #%d (may have been deleted): %w", meta.DeployRequestID, err)
-	}
-
-	return &engine.ControlResult{
-		Accepted:    true,
-		Message:     fmt.Sprintf("Revert window skipped for deploy request #%d", dr.Number),
-		ResumeState: req.ResumeState,
-	}, nil
-}
-
-// controlMeta extracts and validates psMetadata from a control request.
-func (e *Engine) controlMeta(req *engine.ControlRequest) (*psMetadata, error) {
-	if req.ResumeState == nil || req.ResumeState.Metadata == "" {
-		return nil, fmt.Errorf("no active schema change")
-	}
-	meta, err := decodePSMetadata(req.ResumeState.Metadata)
-	if err != nil {
-		return nil, fmt.Errorf("decode resume state: %w", err)
-	}
-	if meta.DeployRequestID == 0 {
-		return nil, fmt.Errorf("no active schema change")
-	}
-	return meta, nil
-}
-
-// --- Volume ---
-
-// Volume adjusts schema change speed by setting the Vitess throttle ratio.
-// Volume 1 = fully throttled (ratio 1.0), Volume 11 = full speed (ratio 0.0).
-// NOTE: Volume/Throttle requires the PlanetScale client to be initialized with a
-// base URL (via Credentials.DSN). This is wired in the tern layer.
-func (e *Engine) Volume(ctx context.Context, req *engine.VolumeRequest) (*engine.VolumeResult, error) {
-	if req.ResumeState == nil || req.ResumeState.Metadata == "" {
-		return nil, fmt.Errorf("no active schema change")
-	}
-	meta, err := decodePSMetadata(req.ResumeState.Metadata)
-	if err != nil {
-		return nil, fmt.Errorf("decode resume state: %w", err)
-	}
-	if meta.DeployRequestID == 0 {
-		return nil, fmt.Errorf("no active schema change")
-	}
-
-	client, err := e.getClient(req.Credentials)
-	if err != nil {
-		return nil, fmt.Errorf("get planetscale client: %w", err)
-	}
-
-	if req.Volume < 1 || req.Volume > 11 {
-		e.logger.Warn("volume out of range, clamping to [1, 11]", "requested", req.Volume)
-	}
-	ratio := volumeToThrottleRatio(req.Volume)
-
-	err = client.ThrottleDeployRequest(ctx, &psclient.ThrottleDeployRequestRequest{
-		Organization:  credOrg(req.Credentials),
-		Database:      req.Database,
-		Number:        meta.DeployRequestID,
-		ThrottleRatio: ratio,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("throttle deploy request: %w", err)
-	}
-
-	return &engine.VolumeResult{
-		Accepted:       true,
-		PreviousVolume: 0, // Unknown — PlanetScale has no query API for current ratio
-		NewVolume:      req.Volume,
-		Message:        fmt.Sprintf("Throttle ratio set to %.0f%%", ratio*100),
-	}, nil
-}
-
-// DefaultVolume is the default throttle volume for new deploys.
-// Maps to a throttle ratio of 0.85 — aggressive enough to limit impact on
-// production traffic while still making progress.
-const DefaultVolume int32 = 2
-
-// volumeToThrottleRatio converts volume (1-11) to a PlanetScale throttle ratio.
-// Lower volume = more throttling. DefaultVolume (2) maps to 0.85.
-// See engine.VolumeRequest for how volume semantics differ between engines.
-var volumeThrottleMap = [12]float64{
-	0:  0.95, // unused (volume is 1-indexed)
-	1:  0.95, // max throttle
-	2:  0.85, // default
-	3:  0.75,
-	4:  0.65,
-	5:  0.55,
-	6:  0.45,
-	7:  0.35,
-	8:  0.25,
-	9:  0.15,
-	10: 0.05,
-	11: 0.0, // no throttle
-}
-
-func volumeToThrottleRatio(volume int32) float64 {
-	if volume <= 1 {
-		return volumeThrottleMap[1]
-	}
-	if volume >= 11 {
-		return volumeThrottleMap[11]
-	}
-	return volumeThrottleMap[volume]
-}
-
-// --- SHOW VITESS_MIGRATIONS shard progress ---
-
-// captureExistingContexts returns the set of migration_context values currently
-// in SHOW VITESS_MIGRATIONS. Used as a baseline before deploying so that new
-// contexts can be identified after deploy.
-func (e *Engine) captureExistingContexts(ctx context.Context, client psclient.PSClient, database string, creds *engine.Credentials) map[string]bool {
-	existing := make(map[string]bool)
-	if creds.DSN == "" {
-		return existing
-	}
-
-	branch := mainBranch(creds)
-	keyspaces, err := client.ListKeyspaces(ctx, &ps.ListKeyspacesRequest{
-		Organization: credOrg(creds),
-		Database:     database,
-		Branch:       branch,
-	})
-	if err != nil {
-		e.logger.Warn("captureExistingContexts: failed to list keyspaces", "error", err)
-		return existing
-	}
-
-	for _, ks := range keyspaces {
-		rows, err := e.showVitessMigrationsForKeyspace(ctx, creds.DSN, ks.Name, "")
-		if err != nil {
-			e.logger.Debug("capture existing contexts: query failed", "keyspace", ks.Name, "error", err)
-			continue
-		}
-		for _, r := range rows {
-			if r.MigrationContext != "" {
-				existing[r.MigrationContext] = true
-			}
-		}
-	}
-
-	e.logger.Info("captured schema change context baseline", "count", len(existing))
-	return existing
-}
-
-// discoverMigrationContext finds the new migration_context that appeared after
-// deploying by comparing current contexts against the pre-deploy baseline.
-func (e *Engine) discoverMigrationContext(ctx context.Context, client psclient.PSClient, database string, creds *engine.Credentials, existingContexts map[string]bool) string {
-	if creds.DSN == "" {
-		e.logger.Debug("skipping schema change context discovery, no DSN configured")
-		return ""
-	}
-
-	e.logger.Info("discovering schema change context", "database", database, "baseline_count", len(existingContexts))
-
-	branch := mainBranch(creds)
-	keyspaces, err := client.ListKeyspaces(ctx, &ps.ListKeyspacesRequest{
-		Organization: credOrg(creds),
-		Database:     database,
-		Branch:       branch,
-	})
-	if err != nil {
-		e.logger.Warn("failed to list keyspaces for schema change context discovery", "error", err)
-		return ""
-	}
-
-	for _, ks := range keyspaces {
-		rows, err := e.showVitessMigrationsForKeyspace(ctx, creds.DSN, ks.Name, "")
-		if err != nil {
-			e.logger.Debug("failed to query schema changes for keyspace", "keyspace", ks.Name, "error", err)
-			continue
-		}
-		for _, r := range rows {
-			if r.MigrationContext != "" && !existingContexts[r.MigrationContext] {
-				e.logger.Info("discovered schema change context", "context", r.MigrationContext)
-				return r.MigrationContext
-			}
-		}
-	}
-
-	e.logger.Warn("schema change context not discovered yet")
-	return ""
-}
-
-// vitessMigrationRow holds a single row from SHOW VITESS_MIGRATIONS.
-type vitessMigrationRow struct {
-	MigrationUUID    string
-	MigrationContext string
-	Keyspace         string
-	Shard            string
-	Table            string
-	Status           string // queued, running, ready_to_complete, complete, failed, cancelled
-	ReadyToComplete  bool
-	DDLAction        string
-	Progress         int
-	ETASeconds       int64
-	RowsCopied       int64
-	TableRows        int64
-	IsImmediate      bool
-	CutoverAttempts  int
-	StartedAt        *time.Time
-	CompletedAt      *time.Time
-}
-
-// queryVitessMigrations queries SHOW VITESS_MIGRATIONS across all keyspaces via vtgate
-// and aggregates per-shard results into per-table TableProgress entries.
-func (e *Engine) queryVitessMigrations(ctx context.Context, client psclient.PSClient, database string, creds *engine.Credentials, migrationContext string) ([]engine.TableProgress, int) {
-	branch := mainBranch(creds)
-	keyspaces, err := client.ListKeyspaces(ctx, &ps.ListKeyspacesRequest{
-		Organization: credOrg(creds),
-		Database:     database,
-		Branch:       branch,
-	})
-	if err != nil {
-		e.logger.Warn("queryVitessMigrations: failed to list keyspaces", "error", err)
-		return nil, 0
-	}
-
-	var allRows []vitessMigrationRow
-	for _, ks := range keyspaces {
-		rows, err := e.showVitessMigrationsForKeyspace(ctx, creds.DSN, ks.Name, migrationContext)
-		if err != nil {
-			e.logger.Error("per-shard progress query failed", "keyspace", ks.Name, "database", database, "error", err)
-			continue
-		}
-		allRows = append(allRows, rows...)
-	}
-
-	if len(allRows) == 0 {
-		return nil, 0
-	}
-
-	return aggregateShardProgress(allRows)
-}
-
-// showVitessMigrationsForKeyspace connects to vtgate and runs
-// SHOW VITESS_MIGRATIONS LIKE '<context>' for a single keyspace.
-// If migrationContext is empty, returns all migrations.
-func (e *Engine) showVitessMigrationsForKeyspace(ctx context.Context, dsn, keyspace, migrationContext string) ([]vitessMigrationRow, error) {
-	if migrationContext != "" {
-		if err := validateMigrationContext(migrationContext); err != nil {
-			return nil, fmt.Errorf("validate context for keyspace %s: %w", keyspace, err)
-		}
-	}
-
-	db, err := e.getVtgateDB(ctx, dsn)
-	if err != nil {
-		return nil, fmt.Errorf("get vtgate connection for keyspace %s: %w", keyspace, err)
-	}
-
-	conn, err := db.Conn(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("get connection: %w", err)
-	}
-	defer utils.CloseAndLog(conn)
-
-	if _, err := conn.ExecContext(ctx, "USE `"+keyspace+"`"); err != nil {
-		return nil, fmt.Errorf("use keyspace %s: %w", keyspace, err)
-	}
-
-	query := "SHOW VITESS_MIGRATIONS"
-	if migrationContext != "" {
-		query += " LIKE '" + migrationContext + "'"
-	}
-	rows, err := conn.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("show vitess_migrations: %w", err)
-	}
-	defer utils.CloseAndLog(rows)
-
-	columns, err := rows.Columns()
-	if err != nil {
-		return nil, fmt.Errorf("get columns: %w", err)
-	}
-
-	var result []vitessMigrationRow
-	for rows.Next() {
-		colValues := make([]sql.NullString, len(columns))
-		colPtrs := make([]any, len(columns))
-		for i := range colValues {
-			colPtrs[i] = &colValues[i]
-		}
-		if err := rows.Scan(colPtrs...); err != nil {
-			e.logger.Debug("scan vitess_migrations row", "keyspace", keyspace, "error", err)
-			continue
-		}
-		colMap := make(map[string]string)
-		for i, col := range columns {
-			if colValues[i].Valid {
-				colMap[col] = colValues[i].String
-			}
-		}
-
-		row := vitessMigrationRow{
-			MigrationUUID:    colMap["migration_uuid"],
-			MigrationContext: colMap["migration_context"],
-			Keyspace:         colMap["keyspace"],
-			Shard:            colMap["shard"],
-			Table:            colMap["mysql_table"],
-			Status:           colMap["migration_status"],
-			ReadyToComplete:  colMap["ready_to_complete"] == "1",
-			DDLAction:        colMap["ddl_action"],
-			IsImmediate:      colMap["is_immediate_operation"] == "1",
-		}
-		if v, err := strconv.Atoi(colMap["progress"]); err != nil && colMap["progress"] != "" {
-			e.logger.Debug("parse vitess_migrations field", "field", "progress", "value", colMap["progress"], "error", err)
-		} else {
-			row.Progress = v
-		}
-		if v, err := parseInt64(colMap["eta_seconds"]); err != nil {
-			e.logger.Debug("parse vitess_migrations field", "field", "eta_seconds", "value", colMap["eta_seconds"], "error", err)
-		} else {
-			row.ETASeconds = v
-		}
-		if v, err := parseInt64(colMap["rows_copied"]); err != nil {
-			e.logger.Debug("parse vitess_migrations field", "field", "rows_copied", "value", colMap["rows_copied"], "error", err)
-		} else {
-			row.RowsCopied = v
-		}
-		if v, err := parseInt64(colMap["table_rows"]); err != nil {
-			e.logger.Debug("parse vitess_migrations field", "field", "table_rows", "value", colMap["table_rows"], "error", err)
-		} else {
-			row.TableRows = v
-		}
-		if v, err := parseInt64(colMap["cutover_attempts"]); err == nil {
-			row.CutoverAttempts = int(v)
-		}
-
-		if ts, parseErr := time.Parse("2006-01-02 15:04:05", colMap["started_timestamp"]); parseErr == nil {
-			row.StartedAt = &ts
-		}
-		if ts, parseErr := time.Parse("2006-01-02 15:04:05", colMap["completed_timestamp"]); parseErr == nil {
-			row.CompletedAt = &ts
-		}
-
-		result = append(result, row)
-	}
-	return result, rows.Err()
-}
-
-// validateMigrationContext rejects migration context strings containing unsafe characters.
-func validateMigrationContext(s string) error {
-	if strings.ContainsAny(s, "'\"\\`") {
-		return fmt.Errorf("invalid context: contains unsafe characters")
-	}
-	return nil
-}
-
-func parseInt64(s string) (int64, error) {
-	if s == "" {
-		return 0, nil
-	}
-	return strconv.ParseInt(s, 10, 64)
-}
-
-// aggregateShardProgress groups SHOW VITESS_MIGRATIONS rows by migration_uuid
-// and produces per-table progress with per-shard breakdown.
-func aggregateShardProgress(rows []vitessMigrationRow) ([]engine.TableProgress, int) {
-	type tableKey struct {
-		keyspace string
-		table    string
-		uuid     string
-	}
-	type shardData struct {
-		shard           string
-		status          string
-		readyToComplete bool
-		progress        int
-		rowsCopied      int64
-		tableRows       int64
-		etaSeconds      int64
-		isImmediate     bool
-		cutoverAttempts int
-		startedAt       *time.Time
-		completedAt     *time.Time
-	}
-
-	tableShards := make(map[tableKey][]shardData)
-	tableOrder := make([]tableKey, 0)
-
-	for _, r := range rows {
-		key := tableKey{keyspace: r.Keyspace, table: r.Table, uuid: r.MigrationUUID}
-		if _, exists := tableShards[key]; !exists {
-			tableOrder = append(tableOrder, key)
-		}
-		tableShards[key] = append(tableShards[key], shardData{
-			shard:           r.Shard,
-			status:          r.Status,
-			readyToComplete: r.ReadyToComplete,
-			progress:        r.Progress,
-			rowsCopied:      r.RowsCopied,
-			tableRows:       r.TableRows,
-			etaSeconds:      r.ETASeconds,
-			isImmediate:     r.IsImmediate,
-			cutoverAttempts: r.CutoverAttempts,
-			startedAt:       r.StartedAt,
-			completedAt:     r.CompletedAt,
-		})
-	}
-
-	var totalRowsCopied, totalTableRows int64
-	var tables []engine.TableProgress
-
-	for _, key := range tableOrder {
-		shards := tableShards[key]
-
-		// Sort shards by Vitess key range for consistent ordering
-		sort.Slice(shards, func(i, j int) bool {
-			return shardLess(shards[i].shard, shards[j].shard)
-		})
-
-		var tblRowsCopied, tblTableRows, maxETA int64
-		var tblProgress int
-		var tblStartedAt *time.Time
-		var latestCompletedAt *time.Time
-		allShardsCompleted := true
-		shardProgress := make([]engine.ShardProgress, len(shards))
-		isInstant := true
-
-		// Determine aggregate table state from shard states
-		tableState := state.Vitess.Complete
-		for i, sh := range shards {
-			tblTableRows += sh.tableRows
-			if sh.etaSeconds > maxETA {
-				maxETA = sh.etaSeconds
-			}
-			if !sh.isImmediate {
-				isInstant = false
-			}
-			// Table started_at = earliest shard started_at
-			if sh.startedAt != nil && (tblStartedAt == nil || sh.startedAt.Before(*tblStartedAt)) {
-				tblStartedAt = sh.startedAt
-			}
-			// Track latest completed_at across shards
-			if sh.completedAt == nil {
-				allShardsCompleted = false
-			} else if latestCompletedAt == nil || sh.completedAt.After(*latestCompletedAt) {
-				latestCompletedAt = sh.completedAt
-			}
-
-			// Resolve effective shard state: running + ready_to_complete = ready_to_complete
-			shardState := sh.status
-			if sh.status == state.Vitess.Running && sh.readyToComplete {
-				shardState = state.Vitess.ReadyToComplete
-			}
-
-			shardPct := sh.progress
-			shardCopied := sh.rowsCopied
-			// When a shard is ready for cutover, the copy phase is complete.
-			// Clamp to 100% since Vitess row counts can lag behind slightly.
-			if shardState == state.Vitess.ReadyToComplete || shardState == state.Vitess.Complete {
-				shardPct = 100
-				if sh.tableRows > 0 && shardCopied < sh.tableRows {
-					shardCopied = sh.tableRows
-				}
-			}
-
-			tblRowsCopied += shardCopied
-
-			shardProgress[i] = engine.ShardProgress{
-				Shard:           sh.shard,
-				State:           shardState,
-				Progress:        shardPct,
-				RowsCopied:      shardCopied,
-				RowsTotal:       sh.tableRows,
-				ETASeconds:      sh.etaSeconds,
-				CutoverAttempts: sh.cutoverAttempts,
-			}
-
-			tableState = resolveTableState(tableState, shardState)
-		}
-
-		if tblTableRows > 0 {
-			tblProgress = int(tblRowsCopied * 100 / tblTableRows)
-		} else if tableState == state.Vitess.Complete || tableState == state.Vitess.ReadyToComplete {
-			tblProgress = 100
-		}
-
-		totalRowsCopied += tblRowsCopied
-		totalTableRows += tblTableRows
-
-		// Table completed_at is only set when all shards have completed.
-		var tblCompletedAt *time.Time
-		if allShardsCompleted {
-			tblCompletedAt = latestCompletedAt
-		}
-
-		tables = append(tables, engine.TableProgress{
-			Namespace:   key.keyspace,
-			Table:       key.table,
-			State:       tableState,
-			Progress:    tblProgress,
-			RowsCopied:  tblRowsCopied,
-			RowsTotal:   tblTableRows,
-			ETASeconds:  maxETA,
-			Shards:      shardProgress,
-			IsInstant:   isInstant,
-			StartedAt:   tblStartedAt,
-			CompletedAt: tblCompletedAt,
-		})
-	}
-
-	overallProgress := 0
-	if totalTableRows > 0 {
-		overallProgress = int(totalRowsCopied * 100 / totalTableRows)
-	} else if len(tables) > 0 {
-		allDone := true
-		for _, t := range tables {
-			if t.State != state.Vitess.Complete && t.State != state.Vitess.ReadyToComplete {
-				allDone = false
-				break
-			}
-		}
-		if allDone {
-			overallProgress = 100
-		}
-	}
-
-	return tables, overallProgress
-}
-
-// resolveTableState merges a shard's state into the current table state.
-// A table has one Vitess migration per shard, each in a different state.
-// This picks the "worst" state so the table reflects the least-progressed shard:
-//
-//	failed            — any shard failed, table is failed
-//	running           — at least one shard still copying rows
-//	queued            — at least one shard not started, none running or failed
-//	ready_to_complete — all shards done copying, waiting for cutover
-//	complete          — all shards finished (initial value)
-func resolveTableState(tableState, shardState string) string {
-	switch shardState {
-	case state.Vitess.Failed, state.Vitess.Cancelled:
-		return state.Vitess.Failed
-	case state.Vitess.Running:
-		if tableState != state.Vitess.Failed {
-			return state.Vitess.Running
-		}
-	case state.Vitess.Queued, state.Vitess.Ready, state.Vitess.Requested:
-		if tableState != state.Vitess.Failed && tableState != state.Vitess.Running {
-			return state.Vitess.Queued
-		}
-	case state.Vitess.ReadyToComplete:
-		if tableState == state.Vitess.Complete {
-			return state.Vitess.ReadyToComplete
-		}
-	}
-	return tableState
-}
-
-// shardLess compares two Vitess shard key ranges for sorting.
-func shardLess(a, b string) bool {
-	aStart := ""
-	bStart := ""
-	if idx := strings.Index(a, "-"); idx > 0 {
-		aStart = a[:idx]
-	}
-	if idx := strings.Index(b, "-"); idx > 0 {
-		bStart = b[:idx]
-	}
-	if aStart == "" && bStart != "" {
-		return true
-	}
-	if aStart != "" && bStart == "" {
-		return false
-	}
-	return aStart < bStart
-}
-
-// verifyBranchMatchesDesiredWithRetry retries verifyBranchMatchesDesired up to
-// 90s to handle PlanetScale VSchema API staleness. DDL schema is fetched via
-// MySQL (real-time) and fails fast on mismatch. Only VSchema errors are
-// retried, since GetKeyspaceVSchema may return stale data after
-// UpdateKeyspaceVSchema.
-func (e *Engine) verifyBranchMatchesDesiredWithRetry(ctx context.Context, client psclient.PSClient, org, database, branch string, keyspaces []string, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword) error {
-	const maxAttempts = 18
-	const pollInterval = 5 * time.Second
-
-	var lastErr error
-	for attempt := range maxAttempts {
-		if attempt > 0 {
-			e.logger.Info("retrying branch schema validation, waiting for VSchema API to converge",
-				"branch", branch, "attempt", attempt+1, "delay", pollInterval)
-			select {
-			case <-ctx.Done():
-				return fmt.Errorf("context cancelled waiting for schema validation: %w", ctx.Err())
-			case <-time.After(pollInterval):
-			}
-		}
-
-		lastErr = e.verifyBranchMatchesDesired(ctx, client, org, database, branch, keyspaces, schemaFiles, password)
-		if lastErr == nil {
-			if attempt > 0 {
-				e.logger.Info("branch schema validated after retry",
-					"branch", branch, "attempts", attempt+1)
-			}
-			return nil
-		}
-
-		// Only retry VSchema staleness errors. DDL validation uses MySQL
-		// (real-time) so DDL mismatches are genuine failures.
-		if !strings.Contains(lastErr.Error(), "unexpected VSchema difference") {
-			return lastErr
-		}
-		e.logger.Debug("VSchema validation attempt failed (API may be stale)",
-			"branch", branch, "attempt", attempt+1, "error", lastErr)
-	}
-	return fmt.Errorf("VSchema still mismatched after %d attempts (%ds): %w",
-		maxAttempts, maxAttempts*int(pollInterval.Seconds()), lastErr)
-}
-
-// verifyBranchMatchesDesired validates that the branch schema matches the
-// desired schema files for all keyspaces. Fetches DDL schema directly via
-// MySQL (LoadSchemaFromDB) to avoid PlanetScale's GetBranchSchema API, which
-// returns stale schema until an asynchronous schema snapshot completes after
-// DDL execution.
-func (e *Engine) verifyBranchMatchesDesired(ctx context.Context, client psclient.PSClient, org, database, branch string, keyspaces []string, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword) error {
-	branchSchema, err := e.fetchBranchSchemaViaMySQL(ctx, password, keyspaces)
-	if err != nil {
-		return fmt.Errorf("fetch branch schema via MySQL for validation: %w", err)
-	}
-
-	for _, ks := range keyspaces {
-		ns := schemaFiles[ks]
-		if ns == nil {
-			continue
-		}
-
-		ddlChanges, vschemaChanged, err := e.diffKeyspace(ctx, client, org, database, branch, ks, ns, branchSchema)
-		if err != nil {
-			return fmt.Errorf("validate keyspace %s: %w", ks, err)
-		}
-
-		if len(ddlChanges) > 0 {
-			var summaries []string
-			var statements []string
-			for _, ch := range ddlChanges {
-				classified, _ := statement.Classify(ch.DDL)
-				summary := ch.DDL
-				if len(classified) > 0 {
-					summary = fmt.Sprintf("%s %s", classified[0].Type, classified[0].Table)
-				}
-				summaries = append(summaries, summary)
-				statements = append(statements, ch.DDL)
-			}
-			e.logger.Error("branch validation failed: unexpected DDL changes",
-				"keyspace", ks,
-				"branch", branch,
-				"change_count", len(ddlChanges),
-				"changes", summaries,
-				"statements", statements,
-				"branch_table_count", len(branchSchema[ks]),
-				"desired_file_count", len(ns.Files),
-			)
-			return fmt.Errorf("keyspace %s has %d unexpected DDL changes after apply: %v\nstatements:\n%s",
-				ks, len(ddlChanges), summaries, strings.Join(statements, "\n"))
-		}
-
-		if vschemaChanged {
-			return fmt.Errorf("keyspace %s has unexpected VSchema difference after apply — VSchema may not have been applied to the branch", ks)
-		}
-	}
-
-	e.logger.Info("branch schema validated via MySQL — matches desired state",
-		"branch", branch,
-		"keyspaces", len(keyspaces),
-	)
-	return nil
-}
-
-// fetchBranchSchemaViaMySQL connects to the branch via MySQL using the branch
-// password and loads table schemas with LoadSchemaFromDB. This returns the
-// real-time schema, bypassing PlanetScale's cached GetBranchSchema API.
-func (e *Engine) fetchBranchSchemaViaMySQL(ctx context.Context, password *ps.DatabaseBranchPassword, keyspaces []string) (map[string][]table.TableSchema, error) {
-	mysqlCfg := mysql.NewConfig()
-	mysqlCfg.User = password.Username
-	mysqlCfg.Passwd = password.PlainText
-	mysqlCfg.Net = "tcp"
-	mysqlCfg.Addr = password.Hostname
-	mysqlCfg.InterpolateParams = true
-	if mtlsRegistered.Load() {
-		mysqlCfg.TLSConfig = mtlsConfigName
-	}
-
-	var mu sync.Mutex
-	result := make(map[string][]table.TableSchema, len(keyspaces))
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(20)
-
-	for _, keyspace := range keyspaces {
-		ks := keyspace
-		g.Go(func() error {
-			ksCfg := mysqlCfg.Clone()
-			ksCfg.DBName = ks
-			db, err := sql.Open("mysql", ksCfg.FormatDSN())
-			if err != nil {
-				return fmt.Errorf("open branch MySQL for keyspace %s: %w", ks, err)
-			}
-			defer utils.CloseAndLog(db)
-
-			if err := db.PingContext(gCtx); err != nil {
-				return fmt.Errorf("ping branch MySQL for keyspace %s: %w", ks, err)
-			}
-
-			tables, err := table.LoadSchemaFromDB(gCtx, db, table.WithoutUnderscoreTables)
-			if err != nil {
-				return fmt.Errorf("load schema for keyspace %s: %w", ks, err)
-			}
-			mu.Lock()
-			result[ks] = tables
-			mu.Unlock()
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-// diffKeyspace diffs a single keyspace's schema between a branch and the
-// desired schema files. Returns DDL changes and whether VSchema differs.
-// Shared by Plan() and verifyBranchMatchesDesired().
-func (e *Engine) diffKeyspace(ctx context.Context, client psclient.PSClient, org, database, branch, ks string, ns *schema.Namespace, currentSchema map[string][]table.TableSchema) ([]engine.TableChange, bool, error) {
-	var currentTableSchemas []table.TableSchema
-	if tables, ok := currentSchema[ks]; ok {
-		currentTableSchemas = append(currentTableSchemas, tables...)
-	}
-
-	desiredTableSchemas, parseErr := parseDesiredSchemas(ks, ns)
-	if parseErr != nil {
-		return nil, false, parseErr
-	}
-
-	plan, planErr := lint.PlanChanges(currentTableSchemas, desiredTableSchemas, nil, e.linter.SpiritConfig())
-	if planErr != nil {
-		return nil, false, fmt.Errorf("plan changes for keyspace %s: %w", ks, planErr)
-	}
-
-	if len(plan.Changes) > 0 {
-		e.logger.Info("diffKeyspace: changes detected",
-			"keyspace", ks,
-			"change_count", len(plan.Changes),
-			"current_table_count", len(currentTableSchemas),
-			"desired_table_count", len(desiredTableSchemas),
-		)
-		for _, pc := range plan.Changes {
-			e.logger.Info("diffKeyspace: change detail",
-				"keyspace", ks,
-				"table", pc.TableName,
-				"statement", pc.Statement[:min(len(pc.Statement), 200)],
-			)
-		}
-		// Log table names from both sides for debugging
-		var currentNames, desiredNames []string
-		for _, t := range currentTableSchemas {
-			currentNames = append(currentNames, t.Name)
-		}
-		for _, t := range desiredTableSchemas {
-			desiredNames = append(desiredNames, t.Name)
-		}
-		e.logger.Info("diffKeyspace: table names",
-			"keyspace", ks,
-			"current", currentNames,
-			"desired", desiredNames,
-		)
-	}
-
-	var tableChanges []engine.TableChange
-	for _, pc := range plan.Changes {
-		stmtType, _, classifyErr := ddl.ClassifyStatement(pc.Statement)
-		if classifyErr != nil {
-			return nil, false, fmt.Errorf("classify statement in keyspace %s: %w", ks, classifyErr)
-		}
-		change := engine.TableChange{
-			Table:     pc.TableName,
-			Operation: stmtType,
-			DDL:       pc.Statement,
-		}
-		if errViolations := pc.Errors(); len(errViolations) > 0 {
-			change.IsUnsafe = true
-			msgs := make([]string, len(errViolations))
-			for i, v := range errViolations {
-				msgs[i] = v.Message
-			}
-			change.UnsafeReason = strings.Join(msgs, "; ")
-		}
-		tableChanges = append(tableChanges, change)
-	}
-
-	// Check VSchema diff
-	vschemaChanged := false
-	if content, ok := ns.Files["vschema.json"]; ok && content != "" {
-		currentVSchema, fetchErr := client.GetKeyspaceVSchema(ctx, &ps.GetKeyspaceVSchemaRequest{
-			Organization: org,
-			Database:     database,
-			Branch:       branch,
-			Keyspace:     ks,
-		})
-		if fetchErr != nil {
-			e.logger.Warn("failed to fetch VSchema for diff, treating as empty",
-				"keyspace", ks, "error", fetchErr)
-		}
-		currentVSchemaRaw := ""
-		if currentVSchema != nil {
-			currentVSchemaRaw = currentVSchema.Raw
-		}
-		vschemaChanged = VSchemaChanged(currentVSchemaRaw, content)
-		if vschemaChanged {
-			e.logger.Info("diffKeyspace: VSchema mismatch detected",
-				"keyspace", ks,
-				"branch", branch,
-				"current_normalized", normalizeVSchemaJSON(currentVSchemaRaw),
-				"desired_normalized", normalizeVSchemaJSON(content),
-			)
-		}
-	}
-
-	return tableChanges, vschemaChanged, nil
-}
-
-// verifyBranchMatchesMain uses Spirit's differ to compare the branch schema
-// against main for the given keyspaces. Returns an error if any DDL changes
-// exist, indicating the branch has stale DDL from a previous failed apply
-// that RefreshSchema did not clean up.
-func (e *Engine) verifyBranchMatchesMain(ctx context.Context, client psclient.PSClient, org, database, branchName, mainBranch string, keyspaces []string, schemaFiles schema.SchemaFiles, password *ps.DatabaseBranchPassword) error {
-	// Fetch main schema via API (stable, not recently modified) and branch
-	// schema via MySQL (real-time, avoids API staleness after RefreshSchema).
-	mainSchema, err := e.fetchDatabaseSchema(ctx, client, org, database, mainBranch, keyspaces)
-	if err != nil {
-		return fmt.Errorf("fetch main schema: %w", err)
-	}
-	branchSchema, err := e.fetchBranchSchemaViaMySQL(ctx, password, keyspaces)
-	if err != nil {
-		return fmt.Errorf("fetch branch schema via MySQL: %w", err)
-	}
-
-	// Build a Namespace from main's schema so we can use diffKeyspace.
-	// This diffs branch (current) against main (desired) — any changes
-	// mean the branch is ahead of main.
-	for _, ks := range keyspaces {
-		mainTables := mainSchema[ks]
-		branchTables := branchSchema[ks]
-
-		// Quick length check — different table counts means mismatch
-		if len(branchTables) != len(mainTables) {
-			return fmt.Errorf("keyspace %s: branch has %d tables, main has %d — branch has stale state from a previous apply",
-				ks, len(branchTables), len(mainTables))
-		}
-
-		// Use Spirit to diff branch vs main (normalized comparison)
-		mainNS := &schema.Namespace{Files: make(map[string]string)}
-		for _, t := range mainTables {
-			mainNS.Files[t.Name+".sql"] = t.Schema + ";"
-		}
-
-		changes, _, diffErr := e.diffKeyspace(ctx, client, org, database, branchName, ks, mainNS, branchSchema)
-		if diffErr != nil {
-			return fmt.Errorf("diff branch vs main for %s: %w", ks, diffErr)
-		}
-		if len(changes) > 0 {
-			return fmt.Errorf("keyspace %s: branch has %d DDL differences from main after refresh — branch has stale state from a previous apply",
-				ks, len(changes))
-		}
-	}
-
-	e.logger.Info("branch schema matches main", "branch", branchName, "keyspaces", len(keyspaces))
-	return nil
-}
-
-// --- Helper functions ---
-
-func (e *Engine) fetchDatabaseSchema(ctx context.Context, client psclient.PSClient, org, database, branch string, keyspaces []string) (map[string][]table.TableSchema, error) {
-	var mu sync.Mutex
-	result := make(map[string][]table.TableSchema, len(keyspaces))
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(20)
-
-	for _, keyspace := range keyspaces {
-		ks := keyspace
-		g.Go(func() error {
-			schemaResult, err := client.GetBranchSchema(gCtx, &ps.BranchSchemaRequest{
-				Organization: org,
-				Database:     database,
-				Branch:       branch,
-				Keyspace:     ks,
-			})
-			if err != nil {
-				var psErr *ps.Error
-				if errors.As(err, &psErr) && psErr.Code == ps.ErrNotFound {
-					// Keyspace doesn't exist yet — treat as empty so all
-					// tables appear as CREATEs in the diff.
-					e.logger.Info("keyspace not found on branch, treating as empty",
-						"keyspace", ks, "branch", branch)
-					mu.Lock()
-					result[ks] = nil
-					mu.Unlock()
-					return nil
-				}
-				return fmt.Errorf("fetch schema for keyspace %s: %w", ks, err)
-			}
-
-			tables := make([]table.TableSchema, len(schemaResult))
-			for i, t := range schemaResult {
-				tables[i] = table.TableSchema{Name: t.Name, Schema: t.Raw}
-			}
-			mu.Lock()
-			result[ks] = tables
-			mu.Unlock()
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (e *Engine) fetchPlanSchema(ctx context.Context, client psclient.PSClient, org, database, branch string, creds *engine.Credentials, keyspaces []string) (map[string][]table.TableSchema, error) {
-	parent, err := client.GetBranch(ctx, &ps.GetDatabaseBranchRequest{
-		Organization: org,
-		Database:     database,
-		Branch:       branch,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("get branch %s: %w", branch, err)
-	}
-
-	if parent.SafeMigrations {
-		e.logger.Debug("using PlanetScale schema API for plan", "database", database, "branch", branch)
-		return e.fetchDatabaseSchema(ctx, client, org, database, branch, keyspaces)
-	}
-
-	if creds == nil || creds.DSN == "" {
-		return nil, fmt.Errorf("safe schema changes are not enabled on branch %q of database %q and vtgate DSN is not configured", branch, database)
-	}
-
-	e.logger.Info("using vtgate schema for plan because PlanetScale safe schema changes are disabled", "database", database, "branch", branch)
-	return e.fetchVtgateSchema(ctx, creds.DSN, keyspaces)
-}
-
-func (e *Engine) fetchVtgateSchema(ctx context.Context, dsn string, keyspaces []string) (map[string][]table.TableSchema, error) {
-	var mu sync.Mutex
-	result := make(map[string][]table.TableSchema, len(keyspaces))
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(20)
-
-	for _, keyspace := range keyspaces {
-		ks := keyspace
-		g.Go(func() error {
-			db, err := e.getVtgateKeyspaceDB(gCtx, dsn, ks)
-			if err != nil {
-				return fmt.Errorf("get vtgate connection for keyspace %s: %w", ks, err)
-			}
-			tables, err := table.LoadSchemaFromDB(gCtx, db, table.WithoutUnderscoreTables)
-			if err != nil {
-				return fmt.Errorf("load schema for keyspace %s: %w", ks, err)
-			}
-			mu.Lock()
-			result[ks] = tables
-			mu.Unlock()
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (e *Engine) createBranch(ctx context.Context, client psclient.PSClient, org, database, branchName, parentBranch string) (*ps.DatabaseBranch, error) {
-	getCtx, getCancel := context.WithTimeout(ctx, 10*time.Second)
-	defer getCancel()
-
-	parent, err := client.GetBranch(getCtx, &ps.GetDatabaseBranchRequest{
-		Organization: org,
-		Database:     database,
-		Branch:       parentBranch,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("get parent branch: %w", err)
-	}
-
-	if !parent.SafeMigrations {
-		return nil, fmt.Errorf("safe schema changes not enabled on branch %q of database %q — enable it in the PlanetScale console before running schema changes", parentBranch, database)
-	}
-
-	createCtx, createCancel := context.WithTimeout(ctx, 30*time.Second)
-	defer createCancel()
-
-	branch, err := client.CreateBranch(createCtx, &ps.CreateDatabaseBranchRequest{
-		Organization: org,
-		Database:     database,
-		Name:         branchName,
-		ParentBranch: parentBranch,
-		Region:       parent.Region.Slug,
-	})
-	if err != nil {
-		// Idempotent: if branch exists, return it
-		if strings.Contains(err.Error(), "Name has already been taken") {
-			e.logger.Info("branch already exists, reusing", "branch", branchName)
-			return client.GetBranch(ctx, &ps.GetDatabaseBranchRequest{
-				Organization: org,
-				Database:     database,
-				Branch:       branchName,
-			})
-		}
-		return nil, fmt.Errorf("create branch %s: %w", branchName, err)
-	}
-	return branch, nil
-}
-
-func (e *Engine) waitForBranchReady(ctx context.Context, client psclient.PSClient, org, database, branchName string) error {
-	ctx, cancel := context.WithTimeout(ctx, 30*time.Minute)
-	defer cancel()
-
-	ticker := time.NewTicker(2 * time.Second)
-	defer ticker.Stop()
-
-	var consecutiveErrors int
-	for {
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("timeout waiting for branch %s", branchName)
-		case <-ticker.C:
-			branch, err := client.GetBranch(ctx, &ps.GetDatabaseBranchRequest{
-				Organization: org,
-				Database:     database,
-				Branch:       branchName,
-			})
-			if err != nil {
-				consecutiveErrors++
-				e.logger.Warn("error checking branch status",
-					"branch", branchName, "error", err, "consecutive_errors", consecutiveErrors)
-				if consecutiveErrors >= 5 {
-					return fmt.Errorf("branch %s not reachable after %d attempts: %w", branchName, consecutiveErrors, err)
-				}
-				continue
-			}
-			consecutiveErrors = 0
-			if branch.Ready {
-				return nil
-			}
-		}
-	}
-}
-
-func (e *Engine) createDeployRequest(ctx context.Context, client psclient.PSClient, org, database, branchName, intoBranch string, autoCutover, autoDeleteBranch bool) (*ps.DeployRequest, error) {
-	return client.CreateDeployRequest(ctx, &ps.CreateDeployRequestRequest{
-		Organization:     org,
-		Database:         database,
-		Branch:           branchName,
-		IntoBranch:       intoBranch,
-		AutoCutover:      autoCutover,
-		AutoDeleteBranch: autoDeleteBranch,
-	})
-}
-
-func (e *Engine) getDeployRequest(ctx context.Context, client psclient.PSClient, org, database string, number uint64) (*ps.DeployRequest, error) {
-	return client.GetDeployRequest(ctx, &ps.GetDeployRequestRequest{
-		Organization: org,
-		Database:     database,
-		Number:       number,
-	})
-}
-
 func generateBranchName(database, planID string) string {
 	sanitized := strings.ReplaceAll(database, "_", "-")
 	if len(sanitized) > 20 {
@@ -3054,8 +714,6 @@ func generateBranchName(database, planID string) string {
 	}
 	return fmt.Sprintf("schemabot-%s-%s", sanitized, shortID)
 }
-
-// --- Deploy state mapping ---
 
 func deployStateToEngineState(drState string) engine.State {
 	switch drState {
@@ -3127,38 +785,6 @@ func deployStateToMessage(drState string) string {
 	default:
 		return fmt.Sprintf("Processing (%s)", drState)
 	}
-}
-
-// --- SQL helpers ---
-
-// parseDesiredSchemas parses CREATE TABLE statements from schema files in a namespace,
-// returning table schemas suitable for diffing against current state. Skips vschema.json
-// and non-.sql files.
-func parseDesiredSchemas(keyspace string, ns *schema.Namespace) ([]table.TableSchema, error) {
-	var schemas []table.TableSchema
-	for filename, content := range ns.Files {
-		if filename == "vschema.json" || !strings.HasSuffix(filename, ".sql") {
-			continue
-		}
-		stmts, err := ddl.SplitStatements(content)
-		if err != nil {
-			return nil, fmt.Errorf("split SQL for keyspace %s: %w", keyspace, err)
-		}
-		for _, stmt := range stmts {
-			ct, err := statement.ParseCreateTable(stmt)
-			if err != nil {
-				return nil, fmt.Errorf("parse desired schema in keyspace %s/%s: %w", keyspace, filename, err)
-			}
-			if err := ddl.ValidateCreateTable(ct); err != nil {
-				return nil, fmt.Errorf("SQL usage error in keyspace %s/%s: %w", keyspace, filename, err)
-			}
-			schemas = append(schemas, table.TableSchema{
-				Name:   ct.TableName,
-				Schema: stmt,
-			})
-		}
-	}
-	return schemas, nil
 }
 
 // sortedKeyspaces returns keyspace names from SchemaFiles in sorted order.
